refactor(ui): consolidate batch command base classes into BatchCommandBase
Merge BatchStatusChangeCommandBase and BatchConfirmationCommandBase into
a single BatchCommandBase with optional confirmation via get_confirmation_config().
