+++ packages/taskdog-ui/src/taskdog/tui/commands/hard_delete.py
@@ -1,31 +1,23 @@
 """Hard delete task command for TUI."""
 
-from taskdog.tui.commands.batch_confirmation_base import BatchConfirmationCommandBase
+from taskdog.tui.commands.batch_command_base import BatchCommandBase
 
 
-class HardDeleteCommand(BatchConfirmationCommandBase):
+class HardDeleteCommand(BatchCommandBase):
     """Command to permanently delete selected task(s) (hard delete)."""
 
-    def get_confirmation_title(self) -> str:
-        """Return the confirmation dialog title."""
-        return "WARNING: PERMANENT DELETION"
-
-    def get_single_task_confirmation(self) -> str:
-        """Return confirmation message for single task."""
+    def get_confirmation_config(self) -> tuple[str, str, str]:
+        """Return confirmation config for permanent deletion."""
         return (
+            "WARNING: PERMANENT DELETION",
             "Are you sure you want to PERMANENTLY delete this task?\n\n"
             "[!] This action CANNOT be undone!\n"
-            "[!] The task will be completely removed from the database."
-        )
-
-    def get_multiple_tasks_confirmation_template(self) -> str:
-        """Return confirmation message template for multiple tasks."""
-        return (
+            "[!] The task will be completely removed from the database.",
             "Are you sure you want to PERMANENTLY delete {count} tasks?\n\n"
             "[!] This action CANNOT be undone!\n"
-            "[!] All tasks will be completely removed from the database."
+            "[!] All tasks will be completely removed from the database.",
         )
 
-    def execute_confirmed_action(self, task_id: int) -> None:
+    def execute_single_task(self, task_id: int) -> None:
         """Permanently delete the task (hard delete)."""
         self.context.api_client.remove_task(task_id)
