+++ packages/taskdog-ui/src/taskdog/tui/commands/pause.py
@@ -1,12 +1,11 @@
 """Pause task command for TUI."""
 
-from taskdog.tui.commands.batch_status_change_base import BatchStatusChangeCommandBase
-from taskdog_core.application.dto.task_operation_output import TaskOperationOutput
+from taskdog.tui.commands.batch_command_base import BatchCommandBase
 
 
-class PauseCommand(BatchStatusChangeCommandBase):
+class PauseCommand(BatchCommandBase):
     """Command to pause the selected task(s)."""
 
-    def execute_single_task(self, task_id: int) -> TaskOperationOutput:
+    def execute_single_task(self, task_id: int) -> None:
         """Pause the task via API client."""
-        return self.context.api_client.pause_task(task_id)
+        self.context.api_client.pause_task(task_id)
