+++ packages/taskdog-ui/src/taskdog/tui/commands/reopen.py
@@ -1,23 +1,19 @@
 """Reopen task command for TUI."""
 
-from taskdog.tui.commands.batch_confirmation_base import BatchConfirmationCommandBase
+from taskdog.tui.commands.batch_command_base import BatchCommandBase
 
 
-class ReopenCommand(BatchConfirmationCommandBase):
+class ReopenCommand(BatchCommandBase):
     """Command to reopen completed or canceled task(s) with confirmation."""
 
-    def get_confirmation_title(self) -> str:
-        """Return the confirmation dialog title."""
-        return "Confirm Reopen"
+    def get_confirmation_config(self) -> tuple[str, str, str]:
+        """Return confirmation config for reopen operation."""
+        return (
+            "Confirm Reopen",
+            "Reopen this task?\n\nStatus will be set to: PENDING",
+            "Reopen {count} tasks?\n\nAll will be set to: PENDING",
+        )
 
-    def get_single_task_confirmation(self) -> str:
-        """Return confirmation message for single task."""
-        return "Reopen this task?\n\nStatus will be set to: PENDING"
-
-    def get_multiple_tasks_confirmation_template(self) -> str:
-        """Return confirmation message template for multiple tasks."""
-        return "Reopen {count} tasks?\n\nAll will be set to: PENDING"
-
-    def execute_confirmed_action(self, task_id: int) -> None:
+    def execute_single_task(self, task_id: int) -> None:
         """Reopen the task via API client."""
         self.context.api_client.reopen_task(task_id)
